requirements.txt
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
black==25.1.0
pillow==11.3.0
